requirements.txt
joblib==1.1.0
MonsterLab==1.2.2
pandas==1.4.2
pymongo[srv]==4.0.2
python-dotenv==0.20.0
scikit-learn==1.0.2
jupyter==1.0.0
plotly==5.6.0
fastapi==0.75.1
uvicorn[standard]==0.17.6
gunicorn==20.1.0
